Initializes open scene, checks to see if path and scene is specified
Path initially set to none to allow no path to be specified. If statements added so that if path isn't specified but is in a non-new scene, the path is set to the scene name. If there is no path and the scene is new, a warning is given and nothing else.