Increment save functionality that increments to next available version
Inclusion of increment_save and next_avail_ver defs that help prevent incrementing over a version already made or saving to a file that isn't of the correct type such as tank. It will increment and save, but will jump ahead to a version number of the correct type that isn't taken. If no existing version exists, it returns 1.