Retrieves and sets a default folder path , added maya.cmds, func cleanup
Use cmds to find path of workspace rootDirectory and then set that as a default folder in "scenes". Def separation of _create_folder_ui functionality from create_ui as code was getting to be unwieldy. Create folder ui returns layout which is called and used in create_ui's addLayout (main).